% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19HcqSccs.R
\docType{package}
\name{Covid19HcqSccs-package}
\alias{Covid19HcqSccs}
\alias{Covid19HcqSccs-package}
\title{Covid19HcqSccs: Self-Controlled Case Series Analysis of the Safety of Hydroxychloroquine}
\description{
Hydroxychloroquine is being considered for use in treatment and prophylaxis of COVID-19 in rapid clinical trials across the world. However, the full safety profiles of this drugs is unknown, and the current trials are unlikely to be powered or have sufficient follow-up time to evaluate most safety outcomes. The aim of this OHDSI study is to use existing retrospective data to evaluate the safety of Hydroxychloroquine, using the self-controlled case series (SCCS) design.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
