% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlledCaseSeries.R
\name{runSelfControlledCaseSeries}
\alias{runSelfControlledCaseSeries}
\title{Execute the Self-Controlled Case Series analyses}
\usage{
runSelfControlledCaseSeries(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "cohort",
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outputFolder,
  maxCores
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outcomeDatabaseSchema}{Schema name where the outcome cohorts are stored. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{outcomeTable}{The name of the table in the outcome database schema that holds the outcome cohorts,}

\item{exposureDatabaseSchema}{Schema name where the exposure cohorts are stored. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{exposureTable}{The name of the table in the exposure database schema that holds the exposure cohorts,}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Execute the Self-Controlled Case Series analyses
}
